Yaml File Generator (#786)
* started the generator for yaml files

* got the first generator online

* generated some yaml files via main

* got the Core translated over

* had the generator handle michelson

* added llvm deps

* LLVM generator up and running

Changes the yaml file to hte ne generated versio

* updated the code to work on plonk

* updated change logs and ran the main again

* Easy Pipeline and Pipeline now Done

* Dit the main directories Stack.yaml

* InteractionNet and InteractionNet IR are both up

* added the missing dependecy for the Main Stack Yaml

The Michelson Morley testing library was missing in the last commit

* Lisp Order Non Cyclic ∧ Make File

The file was updated in order to not be cylcic in the defparameters,
so it can be run as a script with no issues

* updated the contributors guide to reflect the new generator

* wrote a document section at the top

* Added Comments on how to make groups and depencies

* Minor indentation fix

* [ci] changes from CI

* Made the sections much Cleaner

The sectioned off bits resemble actual grouping. The previous sections
were promoted to to 3 ;'s to reflect their bigger grouping

* Added a few more lines to the opening comment text

* added a link to the contributoring guide to the generator script

* updated the changelog to lack the V

* Fixed an issue of building

It seems the main directory builds fine, however sub package have
mysteriously stopped building, this is true on develop and on this
branch. I've tried with two separate machines, however it should be
fixed now?

* Updated Changelog

* Made Changelog Present Tense

As per review on #791 I've changed the style to be present tense and
not past tense.

Co-authored-by: drone <noreply@drone>