Loosened types in functions to rely on constraint-based m.
Loosening the types by using constraints keeps us from restricting the
functions too much based on a specific monad. This became clear by a
mandatory use of `lift` (that will be removed in the next
commit) in `mkLam` to make things work. By using the constraints GHC is
now able to infer the correct monad instances, allowing us to remove the
lift in this case.