LLVM lambda term support (#954)
There are still some issues with this implementation:
- Conflicting names for lambda functions: https://github.com/anoma/juvix/issues/993
- No support for higher order functions: https://github.com/anoma/juvix/issues/996
- Very limited primitives: https://github.com/anoma/juvix/issues/997

These will be taken care of in their own PR's.

* Added a very basic environment for locally bound variables.

* [ci] changes from CI

* Updated environment to contain operands instead of names.

* Rewrote mkApp to use AnnTerm instead of Term.

* Improved comments mkApp.

* Removed pattern match on type in mkApp and removed type argument in applyPrim.

* Implemented basic lambda support.

The implementation needs testing and finetuning.

* Added application of variables.

* Modified mkPrim to take types into account.

* Removed environment from mkPrim.

* Extended comments to mkLam.

* Fixed record field module names in Compilation.hs.

* Preliminary apply instance for PrimVal.

The current implementation is nearly a direct copy from the Michelson
backend and needs improvement in the future.

* Fixed wrong fun argument to takeToReturn call in apply.

* Added function types to typeToLLVM.

* Added globalRef for making global references.

* Added TODO in mkParam.

mkParam now creates parameters with the name "UnName 0_0" (including the
quotes !). This is obviously wrong, but there does not seem a good way
to translate a LLVM.Name into an LLVM.ParameterName. We have to look
into that.

* Implemented Env as a Map.

* Implemented proper code generation when main function is a lambda.

* Marginally improved error on unknown variable.

* Implemented argument passing to applications of lambda's.

* Removed constructor and spaces in parameter names for lambda's.

* Fixed return type of main.

Initially, the whole type of the term was used as the return type.
Obviously this was incorrect, so now the last element of the type is
used as a return type.

* Loosened types in functions to rely on constraint-based m.

Loosening the types by using constraints keeps us from restricting the
functions too much based on a specific monad. This became clear by a
mandatory use of `lift` (that will be removed in the next
commit) in `mkLam` to make things work. By using the constraints GHC is
now able to infer the correct monad instances, allowing us to remove the
lift in this case.

* Removed call to `lift` in `mkLam`.

Prior to redefining the types of the functions using constraints, the
call to lift was mandatory to make things compile. This had the effect
that nested calls of mkLam all created calls at the same level, not
nested.

* Added test files for main lambda.

* Added test files for main apply.

* Correctly pass arguments to call in mkApp.

* Split mkMain from compileProgram.

* Forced environment to be empty for non-lambda main function.

* Wrote elaborate comment for mkMain.

* Wrote small comment for compileProgram.

* Added parameter comment for compileProgram.

* Added parameter comment for mkMain.

* Changed order of functions in Compilation.hs.

* Added small comment to Env.

* Tiny refactoring of mkMain.

* Added comments to compileTerm.

* Added Env comment to mkLam.

* Fixed Env comment in mkApp.

* Added comments to applyPrim.

* Removed commented out code in mkParameterName.

* Added some TODO for future releases.

* Clarified implementation of typeToLLVM a bit.

* Revert "Preliminary apply instance for PrimVal."

This reverts commit 5dc471b89477ac1614d405e06bf86931e8f5bd92.

As the implementation of apply is far from finished, and there is not
implementation yet for interpreting of LLVM code, it is best to revert
this commit for now. That way it becomes immediately clear that it has
not been implemented yet.

* Removed redundant imports in Pipeline.hs.

* Added necessary golden files for translate package.

* [ci] changes from CI

* Rewrote local mkParam into global mkParameter and changed mkLam.

The golden files are updated to match the new variables names that are
used.

* Modified `functionTy` to return tuple of arguments and return type.

* Fixed a mistake in the previous commit.

* Added .hr .ir and .erased golden files for MainApply and MainLambda.

These files are necessary to pass the new tests that got merged from the
develop branch.

* [ci] changes from CI

* Added run-llvm script.

The script allows one to run the Juvix compiler using the LLVM backend,
and immediately translate the result to WASM and execute it.

* Settled on ty for type in variables.

* Renamed functionTy -> functionType.

* Removed superfluous type information in `compileProgram`.

* Fixed run-llvm tmp file leakage.

* Made run-llvm exit earlier on compile errors.

* Removed [[ and ]] bashism in run-llvm.

Co-authored-by: drone <noreply@drone>