Give core tests short names
Currently, the test output for core is almost unreadable, with most tests having "names" consisting of a derived `Show` call spilling onto several lines. Due to the way tasty formats its output, this results in tests which happen to have shorter names having several lines' worth of space characters before them. Overall this leads to the names being impenetrable blobs and huge amounts of empty space.

This commit adds an extra argument of type `TestName` to functions like `shouldCheck`, which in each case is passed a short mnemonic name. When a test fails, it can be found in the source by searching for its name (the previous labels didn't actually occur in the code anywhere), and the names have been chosen to (hopefully) make it obvious when combined with the subheadings which test is which.